update
save or update from modal

+++ 1-project/app/Http/Controllers/GroupController.php
@@ -37,12 +37,12 @@ class GroupController extends Controller
      * @param  \App\Http\Requests\StoreGroupRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(Request $request, Group $group)
     {
-        $group = new Group;
+		if(!$group)
+			$group = new Group;
 		$group->group_title = $request->group_title;
 		$group->group_number = $request->group_number;
-		//$group->deleted_at = null;
 		
 		$group->save();
 		return redirect()->route('group.index');
@@ -94,9 +94,17 @@ class GroupController extends Controller
      * @param  \App\Models\Group  $group
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Group $group)
+    public function destroy(Group $group, $page)
     {
+		
+		$children = $group->groupChildren;
+		if(count($children) != 0){
+			$route = 'group.'.$page;
+			return redirect()->route($route, ['group'=>$group])->with('error_message', 'Trinti negalima, grupėje yra vaikų');
+		}
+	
+		
         $group->delete();
-		return redirect()->route('group.index');
+		return redirect()->route('group.index')->with('success_message', 'Sėkmingai ištrinta');
     }
 }
